+++ setup.py
@@ -1,7 +1,4 @@
-# -*- coding:utf-8 -*-
 import os
-import sys
-
 
 from setuptools import setup, find_packages
 here = os.path.abspath(os.path.dirname(__file__))
@@ -13,7 +10,6 @@ try:
 except IOError:
     README = CHANGES = ''
 
-
 install_requires = [
     'dictknife[load] >= 0.3.1',
     "prestring",
@@ -21,39 +17,44 @@ install_requires = [
     "magicalimport",
 ]
 
-
-docs_extras = [
-]
+docs_extras = []
 
 tests_require = [
+    "testmarker",
 ]
 
-testing_extras = tests_require + [
-]
+testing_extras = tests_require + []
+try:
+    from testmarker.setupcmd import test
+except ImportError:
+    from setuptools.command.test import test
 
-setup(name='swagger-marshmallow-codegen',
-      version='0.4.0',
-      description='generating marshmallow\'s schema from swagger definition file',
-      long_description=README + '\n\n' + CHANGES,
-      classifiers=[
-          "Programming Language :: Python",
-          "Programming Language :: Python :: Implementation :: CPython",
-      ],
-      keywords='swagger,codegen,marshmallow,code-generation,schema',
-      author="podhmo",
-      author_email="ababjam61@gmail.com",
-      url="https://github.com/podhmo/swagger-marshmallow-codegen",
-      packages=find_packages(exclude=["swagger_marshmallow_codegen.tests"]),
-      include_package_data=True,
-      zip_safe=False,
-      install_requires=install_requires,
-      extras_require={
-          'testing': testing_extras,
-          'docs': docs_extras,
-      },
-      tests_require=tests_require,
-      test_suite="swagger_marshmallow_codegen.tests",
-      entry_points="""
+setup(
+    name='swagger-marshmallow-codegen',
+    version='0.4.0',
+    description='generating marshmallow\'s schema from swagger definition file',
+    long_description=README + '\n\n' + CHANGES,
+    classifiers=[
+        "Programming Language :: Python",
+        "Programming Language :: Python :: Implementation :: CPython",
+    ],
+    keywords='swagger,codegen,marshmallow,code-generation,schema',
+    author="podhmo",
+    author_email="ababjam61@gmail.com",
+    url="https://github.com/podhmo/swagger-marshmallow-codegen",
+    packages=find_packages(exclude=["swagger_marshmallow_codegen.tests"]),
+    include_package_data=True,
+    zip_safe=False,
+    install_requires=install_requires,
+    extras_require={
+        'testing': testing_extras,
+        'docs': docs_extras,
+    },
+    cmdclass={"test": test},
+    tests_require=tests_require,
+    test_suite="swagger_marshmallow_codegen.tests",
+    entry_points="""
       [console_scripts]
 swagger-marshmallow-codegen=swagger_marshmallow_codegen.cmd:main
-""")
+"""
+)
